Add Facebook auth via omniauth
This commit adds all the necessary steps for setting up omniauth with
Facebook.

This includes creating a User model, a Sessions controller and updating
the application controller and routes.

It also updates the application layout to have a sign in/out link along
with the user's name when they are signed in.

The Facebook app ID and secret still need to be manually added in the
initializers/omniauth.rb file.

I followed this guide: https://coderwall.com/p/bsfitw